Custom input system
Add the Input::get_char function to replace getch that handles special
keys.
Add the Input::get_line function to replace getnstr which has some
handling for custom keys as well as giving access to the string as it
gets edited and a history.

Add a new command line option to set the maximum size of input
histories.

The search function now shows results while typing.